fix: entity 수정

+++ src/main/java/com/fivemybab/ittabab/report/command/repository/ReportRepository.java
@@ -1,11 +1,9 @@
 package com.fivemybab.ittabab.report.command.repository;
 
-import com.fivemybab.ittabab.report.command.entity.Report;
+import com.fivemybab.ittabab.report.command.domain.aggregate.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface ReportRepository extends JpaRepository<Report, Integer> {
+public interface ReportRepository extends JpaRepository<Report, Long> {
 }
