+++ src/main/java/com/fivemybab/ittabab/schedule/command/repository/ScheduleRepository.java
@@ -1,7 +1,7 @@
 package com.fivemybab.ittabab.schedule.command.repository;
 
-import com.fivemybab.ittabab.schedule.command.entity.ScheduleInfo;
+import com.fivemybab.ittabab.schedule.command.domain.aggregate.ScheduleInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScheduleRepository extends JpaRepository<ScheduleInfo, Integer> {
+public interface ScheduleRepository extends JpaRepository<ScheduleInfo, Long> {
 }
