+++ src/main/java/com/fivemybab/ittabab/store/command/dto/OrderMenuDTO.java
@@ -9,8 +9,8 @@ import lombok.ToString;
 @ToString
 public class OrderMenuDTO {
 
-    private int orderId;
-    private int reviewId;
-    private int menuId;
+    private Long orderId;
+    private Long reviewId;
+    private Long menuId;
 
 }
